requirements.txt
Django~=4.0.6
django-allauth~=0.51.0
django-filter~=22.1
APScheduler~=3.9.1
python-dotenv~=0.20.0
celery~=5.2.7